+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - (Maintenance) Remove support for RELATED_IMAGE_UBI, RELATED_IMAGE_DATABASE and RELATED_IMAGE_METRICSEXPORTER env vars
 - (Bugfix) Fix numactl options
 - (Maintenance) Bump Go to 1.20.10
+- (Bugfix) Fix ArangoBackup Create Backoff & ArangoBackupPolicy propagation
  
 ## [1.2.33](https://github.com/arangodb/kube-arangodb/tree/1.2.33) (2023-09-27)
 - (Maintenance) Bump golang.org/x/net to v0.13.0